fixed issue in repository

+++ hibernate_OneToMany/src/main/java/lk/harshana/dao/VehicleDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import lk.harshana.model.Vehicle;
 
-public interface VehicleDao extends CrudRepository<Vehicle, Long>{
+public interface VehicleDao extends CrudRepository<Vehicle, Integer>{
 
 }
